Conflicts:
	assembler/spades.py
	assembler/src/debruijn/coverage_based_rr.hpp
	assembler/src/debruijn/detail_coverage.hpp
	assembler/src/debruijn/indices/debruijn_edge_index.hpp
	assembler/src/debruijn/long_read_storage.hpp
	assembler/src/debruijn/loop_filter.hpp
	assembler/src/debruijn/pac_index.hpp
	assembler/src/debruijn/pacbio_aligner.hpp
	assembler/src/debruijn/path_extend/bidirectional_path.hpp
	assembler/src/debruijn/path_extend/extension_chooser.hpp
	assembler/src/debruijn/path_extend/path_extend_launch.hpp
	assembler/src/debruijn/path_extend/pe_resolver.hpp
	assembler/src/debruijn/path_extend/split_graph_pair_info.hpp
	assembler/src/debruijn/repeat_resolving_routine.hpp
	assembler/src/debruijn/sequence_mapper.hpp
	assembler/src/io/library.cpp

+++ assembler/src/debruijn/resolved_pair_info.hpp
@@ -38,7 +38,7 @@ public:
 //		VERIFY(labels_.edge_inclusions.find(old_edge)->second.size() > 0);
 		if (!(labels_.edge_inclusions.find(old_edge)->second.size() > 0)){
 			DEBUG("There are no current copy for old graph edge " << old_edge << " " << old_graph_.str(old_edge));
-			no_current_copies_ids.insert(old_graph_.int_id(old_edge));
+			no_current_copies_ids.insert((int) old_graph_.int_id(old_edge));
 			no_current_copies++;
 		}
 		return labels_.edge_inclusions.find(old_edge)->second.size() == 1;
@@ -75,7 +75,7 @@ public:
 			pair<EdgeId, size_t> new_pos_of_second = OldEdgePositionInNewGraph(e2);
 			for (auto it = infos.begin(); it != infos.end(); ++it) {
 				new_pair_info.AddPairInfo(new_pos_of_first.first, new_pos_of_second.first,
-								0. + it->d + new_pos_of_first.second - new_pos_of_second.second,
+								0. + it->d + (double) new_pos_of_first.second - (double) new_pos_of_second.second,
 								it->weight, it->var, false);
 			}
 		}
