+++ assembler/src/include/de/peak_finder.hpp
@@ -364,7 +364,7 @@ private:
       }// else if (j < i && hist_[index_max - x_left_][0] == hist_[j - x_left][0] ) index_max = j;
     TRACE("Maximum is " << index_max);
 
-    if  (abs(index_max - peak) <= delta)
+    if  ((size_t)abs(index_max - peak) <= delta)
       return true;
 
     return false;
